# swagger.yaml
...
swagger: "2.0"
basePath: "/api/v0.2"
info:
  title: LEXIS Portal API
  description: The LEXIS Portal API serves as the API proxy to all the interfaces that interact with the parts involved in LEXIS.
  version: "0.2.2021-12-15"
  contact:
    email: diego@cyclops-labs.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
securityDefinitions:
  APIKeyHeader:
    type: apiKey
    in: header
    name: X-API-KEY
  APIKeyParam:
    type: apiKey
    in: query
    name: api_key
  Keycloak:
    type: oauth2
    flow:  accessCode
    # Need to get clarity on whether these are necessary here or provided with a config file
    authorizationUrl: 'https://138.246.232.245:6443/auth'
    tokenUrl: 'https://138.246.232.245:6443/auth'
security:
  - APIKeyHeader: []
  - APIKeyParam: []
  - Keycloak: []
paths:
...
  /dataset/replicate:
    post:
      tags:
        - dataSetManagement
        - EUDATReplication
      consumes:
        - application/json
      produces:
        - application/json
      summary: This is called when a user requests data to be replicated between different systems. The request will be added to the jobs queue.
      operationId: Replicate
      description: This is called when a user requests data to be replicated between different systems. The request will be added to the jobs queue.
      parameters:
        - in: body
          name: parameters
          description: parameters
          required: true
          schema:
            type: object
            required:
              - source_system
              - source_path
              - target_system
            properties:
              source_system:
                type: string
                description: |
                  Possible values are: "lrz_iRODS", "it4i_iRODS"
              target_system:
                type: string
                description: |
                  Possible values are: "lrz_iRODS", "it4i_iRODS"
              source_path:
                type: string
              target_path:
                type: string
      responses:
        '201':
          description: The response code means that the data transfer has been initiated. Status of the transfer can be checked by querying the status.
          schema:
            $ref: "#/definitions/SteeringRequestID"
        '400':
          description: This means that there's something wrong in the input parameters and the server couldn't understand the request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description:  This means that the user is not authenticated with keycloak and data transfer can't be triggered unless the user first log in with a valid user
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: This means that the resource the user is trying to transfer from or to is not readable or writable by the user. User doesn't have the correct rights to either read the source file or write on the target system location.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description:  This means that either the source path or the the target path on the system doesn't exist.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '414':
          description: This means that the either the source path or the target path is longer than the server is willing to interpret.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '429':
          description: This means that the user has sent too many requests in a given amount of time. The user is advised to wait and send the request again.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: This means that something has gone wrong on the data transfer steering server. The user is advised to wait and send the request again.
          schema:
            $ref: "#/definitions/ErrorResponse"
  /dataset/replicate/{request_id}:
    get:
      tags:
        - dataSetManagement
        - EUDATReplication
      produces:
        - application/json
      summary: This is called when a user requests to check the status of his data replication
      operationId: CheckReplicateStatus
      description: This is called when a user requests to check the status of his data replication
      parameters:
        - in: path
          name: request_id
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: This means that the status has been returned to the user in the response body.
          schema:
            $ref: "#/definitions/DataReplication"
        '400':
          description: This means that there's something wrong in the input parameters and the server couldn't understand the request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description:  This means that the user is not authenticated with keycloak and data transfer can't be triggered unless the user first log in with a valid user
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: This means that the resource the user is trying to transfer from or to is not readable or writable by the user. User doesn't have the correct rights to either read the source file or write on the target system location.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description:  This means that either the source path or the the target path on the system doesn't exist.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '414':
          description: This means that the either the source path or the target path is longer than the server is willing to interpret.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '429':
          description: This means that the user has sent too many requests in a given amount of time. The user is advised to wait and send the request again.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: This means that something has gone wrong on the data transfer steering server. The user is advised to wait and send the request again.
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /dataset/pid/{request_id}:
    get:
      tags:
        - dataSetManagement
        - EUDATReplication
      produces:
        - application/json
      summary: This is called when a user requests to check the status of the PID assignment
      operationId: CheckPIDStatus
      description: This is called when a user requests to check the status of his data replication
      parameters:
        - in: path
          name: request_id
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: This means that the status has been returned to the user in the response body.
          schema:
            $ref: "#/definitions/DataReplication"
        '400':
          description: This means that there's something wrong in the input parameters and the server couldn't understand the request.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '401':
          description:  This means that the user is not authenticated with keycloak and data transfer can't be triggered unless the user first log in with a valid user
          schema:
            $ref: "#/definitions/ErrorResponse"
        '403':
          description: This means that the resource the user is trying to transfer from or to is not readable or writable by the user. User doesn't have the correct rights to either read the source file or write on the target system location.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '404':
          description:  This means that either the source path or the the target path on the system doesn't exist.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '414':
          description: This means that the either the source path or the target path is longer than the server is willing to interpret.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '429':
          description: This means that the user has sent too many requests in a given amount of time. The user is advised to wait and send the request again.
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: This means that something has gone wrong on the data transfer steering server. The user is advised to wait and send the request again.
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  ErrorResponse:
    type: object
    required:
      - errorString
    properties:
      errorString:
        type: string
...
  SteeringRequestID:
    type: object
    required:
      - request_id
    properties:
      request_id:
        type: string
        format: uuid
        example: "31279ed6-e7b1-4c9f-ab26-206c863a2264"
        description: UUID
...
  DataReplication:
    type: object
    required:
      - status
    properties:
      status:
        type: string
        description: Status of the data replication process
        example: "Replication completed"
      PID:
        type: string
        description: EUDAT PID of the replicated data
        example: "1001/e4495e44-dc84-11ea-9c37-0050568f81d2"
      target_path:
        type: string
        description: Path to where the data are replicated
        example: "/IT4ILexisZone/project/proj1fc54ae5e387be0d369eddddc2010e5e/datasetL4988"
...